+++ src/main/java/op/model/ScheduledTask.java
@@ -8,10 +8,10 @@ package op.model;
 public class ScheduledTask {
 
     public Task task;
-    private double startTime;
+    private int startTime;
     private int processor;
 
-    public ScheduledTask (Task task, double startTime, int processor) {
+    public ScheduledTask (Task task, int startTime, int processor) {
         this.task = task;
         this.startTime = startTime;
         this.processor = processor;
